requirements.txt
numpy==2.1.3
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
pandas==2.2.3
Pillow==11.0.0
scikit_learn==1.5.2
streamlit==1.40.2
torch==2.5.1
transformers==4.47.0
ultralytics==8.3.44
